Rename protobuf package to proto
