+++ proto/messages.proto
@@ -17,7 +17,7 @@
 
 syntax = "proto3";
 
-package protobuf;
+package proto;
 
 // A generic message
 message Message {
