+++ server/server.go
@@ -30,7 +30,7 @@ import (
 	"github.com/xlab-si/emmy/config"
 	"github.com/xlab-si/emmy/crypto/groups"
 	"github.com/xlab-si/emmy/log"
-	pb "github.com/xlab-si/emmy/protobuf"
+	pb "github.com/xlab-si/emmy/proto"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials"
 )
